fix issue with changing urls
