+++ src/App.tsx
@@ -2,17 +2,9 @@
 
 import { NavigationContainer } from "@react-navigation/native";
 import { RootStackParamList } from "./Screens/types";
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  from,
-  HttpLink,
-} from "@apollo/client";
+import { ApolloProvider } from "@apollo/client";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-import { RetryLink } from "@apollo/client/link/retry";
 import NoConnection from "@components/NoConnection";
-import { offsetLimitPagination } from "@apollo/client/utilities";
 import { LikeContextProvider } from "./contexts/LikedContext";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import Pokemon from "./Screens/Pokemon";
@@ -28,36 +20,9 @@ import { LightTheme, DarkTheme } from "@constants/themes";
 import * as ExpoLinking from "expo-linking";
 
 import "@notifications/initialize";
+import { client } from "./client";
 
 const prefix = ExpoLinking.createURL("/");
-
-const link = from([
-  new RetryLink({
-    delay: {
-      initial: 300,
-      max: Infinity,
-      jitter: true,
-    },
-    attempts: {
-      max: 5,
-      retryIf: (error, _operation) => !!error,
-    },
-  }),
-  new HttpLink({ uri: "https://beta.pokeapi.co/graphql/v1beta" }),
-]);
-
-export const client = new ApolloClient({
-  link,
-  cache: new InMemoryCache({
-    typePolicies: {
-      Query: {
-        fields: {
-          pokemon_v2_pokemon: offsetLimitPagination(["$ids", "$like", "$name"]),
-        },
-      },
-    },
-  }),
-});
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
