+++ src/notifications/initialize.ts
@@ -1,6 +1,7 @@
 import * as Notifications from "expo-notifications";
 import schedulePokemonOfTheDayNotification from "./triggers/schedulePokemonOfTheDayNotification";
 import registerForPushNotificationsAsync from "./utils/registerForPushNotificationsAsync";
+import setNextRandomPokemonUrl from "./utils/getRandomPokemonUrl";
 
 Notifications.setNotificationHandler({
   handleNotification: async () => ({
@@ -11,6 +12,7 @@ Notifications.setNotificationHandler({
 });
 
 (async () => {
+  await setNextRandomPokemonUrl();
   await registerForPushNotificationsAsync();
   await schedulePokemonOfTheDayNotification();
 })();
